chore(hogql): move generic LLM helper and hog prompts out of hogql/ai.py
posthog/hogql/ai.py accreted a generic OpenAI wrapper (hit_openai) and
~2800 lines of Hog/CDP prompt constants unrelated to HogQL. Move
hit_openai to posthog/llm/completions.py and the CDP prompts to
products/cdp/backend/prompts.py, leaving only the HogQL-coupled
write_sql_from_prompt and its prompts. Pure relocation, no behavior
change.

Generated-By: PostHog Code
Task-Id: 4dba837c-ec3f-4fcc-8dfe-205f8f9f6248
