+++ products/cdp/backend/max_tools.py
@@ -7,29 +7,27 @@ from langchain_core.messages import HumanMessage, SystemMessage
 from pydantic import BaseModel, Field
 
 from posthog.hogql import errors as hogql_errors
-from posthog.hogql.ai import (
-    DESTINATION_LIMITATIONS_MESSAGE,
-    EVENT_PROPERTY_TAXONOMY_MESSAGE,
-    EVENT_TAXONOMY_MESSAGE,
-    FILTER_TAXONOMY_MESSAGE,
-    HOG_EXAMPLE_MESSAGE,
-    HOG_FUNCTION_FILTERS_SYSTEM_PROMPT,
-    HOG_FUNCTION_INPUTS_SYSTEM_PROMPT,
-    HOG_GRAMMAR_MESSAGE,
-    IDENTITY_MESSAGE_HOG,
-    INPUT_SCHEMA_TYPES_MESSAGE,
-    PERSON_TAXONOMY_MESSAGE,
-    TRANSFORMATION_LIMITATIONS_MESSAGE,
-    TRANSFORMATION_STRUCTURE_MESSAGE,
-)
 from posthog.hogql.parser import parse_program
 
 from posthog.cdp.validation import compile_hog
 
 from products.cdp.backend.prompts import (
+    DESTINATION_LIMITATIONS_MESSAGE,
+    EVENT_PROPERTY_TAXONOMY_MESSAGE,
+    EVENT_TAXONOMY_MESSAGE,
+    FILTER_TAXONOMY_MESSAGE,
+    HOG_EXAMPLE_MESSAGE,
     HOG_FUNCTION_FILTERS_ASSISTANT_ROOT_SYSTEM_PROMPT,
+    HOG_FUNCTION_FILTERS_SYSTEM_PROMPT,
     HOG_FUNCTION_INPUTS_ASSISTANT_ROOT_SYSTEM_PROMPT,
+    HOG_FUNCTION_INPUTS_SYSTEM_PROMPT,
+    HOG_GRAMMAR_MESSAGE,
     HOG_TRANSFORMATION_ASSISTANT_ROOT_SYSTEM_PROMPT,
+    IDENTITY_MESSAGE_HOG,
+    INPUT_SCHEMA_TYPES_MESSAGE,
+    PERSON_TAXONOMY_MESSAGE,
+    TRANSFORMATION_LIMITATIONS_MESSAGE,
+    TRANSFORMATION_STRUCTURE_MESSAGE,
 )
 
 from ee.hogai.chat_agent.schema_generator.parsers import PydanticOutputParserException
