+++ products/product_analytics/backend/api/insight_metadata.py
@@ -14,9 +14,8 @@ from posthog.schema import (
     StickinessActorsQuery,
 )
 
-from posthog.hogql.ai import hit_openai
-
 from posthog.event_usage import groups
+from posthog.llm.completions import hit_openai
 from posthog.models import Team
 
 from products.product_analytics.backend.api.ai_billing import billable_ai_properties
